test(workflow-repository): split test_save_semantics.py to keep 500-line rule (Norman R-N1)
Norman R-N1 [却下]: ``test_save_semantics.py`` (502行) が empire-repo PR #29
教訓 (506行→ディレクトリ分割) を踏襲した先で再発した。「対象あり版テンプレート」
として後続 4 件 Repository PR が写経する立場上、+2 行の侵食も許容できないため
クラス境界沿いに 2 ファイルへ分割する。

分割内容:

* ``test_save_delete_then_insert.py`` (251 行)
  - TestSaveDeleteThenInsert  (TC-IT-WFR-009: §確定 B wholesale 置換)
  - TestSaveSqlOrder          (TC-IT-WFR-010: §確定 B 5-step DML order)
  - TestSaveDistinctWorkflowsAreIndependent (workflow_id scoping smoke)
* ``test_save_round_trip.py``        (296 行)
  - TestRolesCsvSortedDeterminism (TC-IT-WFR-011/012: §確定 G sorted CSV)
  - TestRoundTripStructuralEquality (TC-IT-WFR-015: §確定 C round-trip)
  - TestTxBoundaryRespectedByRepository (TC-IT-WFR-016: §確定 B Tx 境界)

両ファイルとも独立した module docstring + import + ``pytestmark`` を持ち、
500 行ルール内で Norman 提案の ≒220/≒280 配分に準拠 (実測 251/296)。
test_save_semantics.py は git rename で round_trip 側に追跡される。

品質: ruff clean、pyright strict 0/0、pytest 593 passed (workflow_repository
27 件 + 既存全テスト無回帰)。

ファイル一覧 (全て 500 行未満、test_workflow_repository ディレクトリ):
* __init__.py                          22
* test_constraints_arch.py            262
* test_masking.py                     207
* test_protocol_crud.py               382
* test_save_delete_then_insert.py     251 (新規)
* test_save_round_trip.py             296 (新規 / 旧 test_save_semantics.py)
